Dedupe lark/weixin channel key helpers
The exact same `normalizeChannelInstanceId`, `channelPlatformKey`, and
`runtimeKey` were defined byte-for-byte in both lark/config.ts and
weixin/config.ts. Move them to channel/shared/channel-keys.ts (the
existing home for cross-channel utilities) and import from there.

Category: b (code reuse). Motivation: pure duplication with no
platform-specific behavior — the kind of drift that already caused
subtle bugs in similar shared helpers in this repo.
